% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdr_add_poly_rgba.R
\name{pdr_add_poly_rgba}
\alias{pdr_add_poly_rgba}
\title{Add an RGBA data set from polygons}
\usage{
pdr_add_poly_rgba(paindrawr_data, invert = FALSE)
}
\description{
Add an RGBA data set from polygons
}
