% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdr_example_file.R
\name{pdr_example_file}
\alias{pdr_example_file}
\title{Local absolute path to example files}
\usage{
pdr_example_file(path = NULL)
}
\arguments{
\item{path}{Typically just the name the file.}
}
\value{
An absolute file path to the specified file.
In case no file is specified, a vector of file names from
the folder, is returned.
}
\description{
This function facilitates users' access to example files
such as e.g. background images, stored in the inst/extdata
folder of the package.
}
\examples{
# Get a vector of files in the _example files_ folder
pdr_example_file()

# Get the absolute path to a file in the _example files_
# folder
pdr_example_data[1,] |> 
  pdr_plot_drawing(background=pdr_example_file("mird_body_background.png"))

}
