% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdr_modify_polygons.R
\name{pdr_modify_polygons}
\alias{pdr_modify_polygons}
\title{Modify pain drawing elements}
\usage{
pdr_modify_polygons(
  paindrawr_data,
  ops = NULL,
  template = NULL,
  polygons_only = FALSE
)
}
\arguments{
\item{paindrawr_data}{The name of the pain drawing list-col in .data}

\item{ops}{A string vector of operations to perform. See below:

Parameter 'ops' can be one of the following, which will be performed the listd order:
\itemize{
\item "make_valid" Removes self-intersection and other issue hich invalidate polygons -- see \code{sf::st_make_valid()}
\item "reduce_to_chull" Reduces polygons to their convex hull -- see \code{sf::st_convex_hull()}
\item "merge_overlaps" Merges (union) overlapping polygons into one -- see \code{sf::st_union()}
}}

\item{.data}{A valid pain drawing tibble}
}
\value{
A valid pain drawing tibble
}
\description{
This function takes a valid pain drawing list as input
and performs one or more operations sequentially on those
data before returning a valid pain drawing list-column of
the same length. It is thus suited for \code{mutate()} operations.
}
\examples{
# Plot gross anatomical regions
pdr_example_anatomy |>
  pdr_implode() |>
  mutate(pdr_data = pdr_polygonize(pdr_data)) |>
  pdr_plot_polygons(pdr_data)

# Plot merged anatomical regions for body outline
pdr_example_anatomy |>
  pdr_implode() |>
  mutate(pdr_data = pdr_polygonize(pdr_data)) |>
  mutate(pdr_data = pdr_modify_polygons(pdr_data, "merge_overlaps")) |>
  pdr_plot_polygons(pdr_data)
}
